fix: move landingePage files into their correct directory.
